# swagger-api.yaml
openapi: 3.0.4
swagger: "2.0"
info:
  title: Swagger Warmhouse - OpenAPI 3.0
  version: 1.0.0
servers:
  - url: http://localhost:8084/api/v1
security:
  - XAuthToken: []
...
paths:
...
  /sensors:
    get:
      tags: [Sensors]
      summary: List sensors
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: house_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by house_id
        - name: type_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by type_id
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by status
      responses:
        "200":
          description: Sensors list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorList"
    post:
      tags: [Sensors]
      summary: Create sensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorCreate"
      responses:
        "201":
          description: Created sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "400":
          $ref: "#/components/responses/BadRequest"
  /sensors/{sensorId}:
    get:
      tags: [Sensors]
      summary: Get sensor by id
      parameters:
        - $ref: "#/components/parameters/SensorId"
      responses:
        "200":
          description: Sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Sensors]
      summary: Update sensor (partial)
      parameters:
        - $ref: "#/components/parameters/SensorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorUpdate"
      responses:
        "200":
          description: Updated sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Sensors]
      summary: Delete sensor
      parameters:
        - $ref: "#/components/parameters/SensorId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /devices:
    get:
      tags: [Devices]
      summary: List devices
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: house_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by house_id
        - name: type_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter by type_id
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by status
      responses:
        "200":
          description: Devices list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceList"
    post:
      tags: [Devices]
      summary: Create device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCreate"
      responses:
        "201":
          description: Created device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
  /devices/{deviceId}:
    get:
      tags: [Devices]
      summary: Get device by id
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Devices]
      summary: Update device (partial)
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Devices]
      summary: Delete device
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /devices/{deviceId}/execute:
    post:
      tags: [Devices]
      summary: Execute device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceExecute"
      responses:
        "200":
          description: Device execute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    XAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Auth token returned by `/auth/login`. Send it in header `X-Auth-Token`.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    SensorId:
      name: sensorId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBadRequest"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorNotFound"
  schemas:
    ErrorBadRequest:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Bad request error"
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    ErrorNotFound:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          example: "NOT FOUND"
        message:
          type: string
          example: "Not found error"
        details:
          type: object
          additionalProperties: true
      required: [code, message]

    # ---------- Users ----------
...
    Sensor:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        house_id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        value:
          type: number
          format: float
          default: 0
        unit:
          type: string
          maxLength: 20
        status:
          type: string
          default: inactive
        last_updated:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
        - id
        - house_id
        - name
        - type_id
        - location
        - serial_number
        - status
        - last_updated
        - created_at
    SensorCreate:
      type: object
      additionalProperties: false
      properties:
        house_id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        value:
          type: number
          format: float
          default: 0
        unit:
          type: string
          maxLength: 20
        status:
          type: string
          default: inactive
      required: [house_id, name, type_id, location, serial_number, status]
    SensorUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        value:
          type: number
          format: float
        unit:
          type: string
          maxLength: 20
        status:
          type: string
      minProperties: 1
    SensorList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Sensor"
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      required: [items, limit, offset, total]

    # ---------- Devices ----------
    Device:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        house_id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        status:
          type: string
          default: inactive
        last_updated:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
        - id
        - house_id
        - name
        - type_id
        - location
        - serial_number
        - status
        - last_updated
        - created_at
    DeviceCreate:
      type: object
      additionalProperties: false
      properties:
        house_id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        status:
          type: string
          default: inactive
      required: [house_id, name, type_id, location, serial_number, status]
    DeviceExecute:
      type: object
      additionalProperties: false
      properties:
        set_value:
          type: integer
          format: int64
      required: []
    DeviceUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 100
        type_id:
          type: integer
          format: int64
        location:
          type: string
          maxLength: 100
        serial_number:
          type: string
          maxLength: 100
        status:
          type: string
      minProperties: 1
    DeviceList:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
      required: [items, limit, offset, total]